requirements.txt
cvzone
opencv-python
numpy
mediapipe
google-generativeai
Pillow
streamlit